Conflicts:
	src/stan/error_handling/matrix/check_ordered.hpp
	src/test/unit/common/initialize_state_test.cpp
	src/test/unit/error_handling/domain_error_test.cpp
	src/test/unit/error_handling/domain_error_vec_test.cpp

+++ src/stan/agrad/fwd/matrix/log_sum_exp.hpp
@@ -29,7 +29,7 @@ namespace stan{
         vals[i] = v(i).val_;
       T deriv(0.0);
       T denominator(0.0);
-      for (size_t i = 0; i < v.size(); ++i) {
+      for (int i = 0; i < v.size(); ++i) {
         T exp_vi = exp(vals[i]);
         denominator += exp_vi;
         deriv += v(i).d_ * exp_vi;
